web/requirements.txt
django>=4.2.23
django-environ>=0.12.0
gunicorn
# whitenoise
django-compressor>=4.5.1
openai
langchain
langchain-openai
langchain-core
langgraph 
redis
django-redis
